pyproject.toml
[build-system]
requires = ["hatchling >= 1.26"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/PythonDMX"]

[project]
name = "PythonDMX"
version = "0.0.1"
authors = [
  { name="Ruben Neurink", email="ruben.neurink@gmail.com" },
]
description = "A DMX512 bit bang over serial implementation"
readme = "README.md"
requires-python = ">=3.9"
classifiers = [
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Operating System :: OS Independent",
]
dependencies = [
    "pyserial"
]
license = "MIT"
license-files = ["LICEN[CS]E*"]

[project.urls]
Homepage = "https://github.com/rneurink/PythonDMX"
Issues = "https://github.com/rneurink/PythonDMX/issues"